moved app js file to root folder
